Make strength reduction RV64-aware
Generalize constant div/rem strength reduction over target width, prioritize power-of-two cases, and allow 64-bit immediates/rematerialization for RV64 magic constants.

Validated with: scripts/run_ir1_qemu_sweep.sh --compiler build/cmd/submission_pipeline --allow-compile-fail overflow

+++ src/riscv/asm_emit.hpp
@@ -14,7 +14,7 @@ std::pair<std::int32_t, std::int32_t> split_imm32(std::int32_t value);
 
 void emit_move(std::vector<AsmInst>& out, PhysicalRegister dest, PhysicalRegister src);
 void emit_symbol_call(std::vector<AsmInst>& out, std::string_view symbol);
-void emit_li(std::vector<AsmInst>& out, PhysicalRegister dest, std::int32_t value);
+void emit_li(std::vector<AsmInst>& out, PhysicalRegister dest, std::int64_t value);
 void emit_add_imm(std::vector<AsmInst>& out,
                   PhysicalRegister dest,
                   PhysicalRegister base,
